A lot more work on the weapon system + new AI health script Weapon interactions work fine, and so does the bullet instantiation Created a file for bullet behaviour, and so far all it will do is destroy itself after a time customizable in the editor Developers in the future will have to customize each different bullet prefab for different destroy times lol
Also started adapting some older weapon interactables to the current system

All in all, I think this is going quite well

+++ Assets/Universal/Scripts/AI/AIHealth.cs
@@ -1,14 +1,13 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 public class AIHealth : MonoBehaviour
 {
-    public float HitPoints = 100;
+    public int HitPoints = 100;
     public ParticleSystem DestructionParticles;
+
     private void Update()
     {
-        if(HitPoints <= 0)
+        if (HitPoints <= 0)
         {
             if (DestructionParticles != null)
             {
